Add QmlComponent::status Move unit tests into tests module. Fix unit tets.
