+++ qmetaobject/Cargo.toml
@@ -25,6 +25,7 @@ cpp_build = "0.5"
 [dev-dependencies]
 cstr = "0.1"
 if_rust_version = "1"
+tempfile = "^3"
 
 
 [package.metadata.docs.rs]
